Sync with upstream

+++ tests/components/test_detector.py
@@ -2,6 +2,7 @@ from sequence.components.detector import *
 from sequence.components.photon import Photon
 from sequence.kernel.timeline import Timeline
 from sequence.utils.encoding import polarization, time_bin
+import numpy as np
 
 
 def create_detector(efficiency=0.9, dark_count=0, count_rate=25e6, time_resolution=150):
@@ -61,7 +62,7 @@ def test_Detector_get():
     # time_resolution
     time_resolution = 233
     detector, parent, tl = create_detector(efficiency=1, count_rate=1e12, time_resolution=time_resolution)
-    times = random.randint(0, 1e12, 100)
+    times = random.randint(0, 1e12, 100, dtype=np.int64)
     times.sort()
     for t in times:
         tl.time = t
